Update windows-template-customization.md
After following this steps have lost acces to the network. Making research have found that "IP Helper" service, who is needed, is dependant on "DHCP" service. Maybe isnt so in every version of Windows, but at lest a side note would be useful for those that face this problem

Thank you